Fix UCRT build and add support for 3.2
Couple of odd things going on.

First, the GCC for UCRT is named the same as in MingW. This was an issue with mini_portile who used the target host for directories. However, the two compiler yield two very different binaries, so I added a new mechanism based on `target_gem_platform` for saving the ports / final gem.

Spend also some time to debug the UCRT build (always changed into `x86_64-linux` when trying to build `tiny_tds`, but adding Ruby 3.2 helped). So I assume it's a big with rake compiler?